Renamed shared_ui to widget, translate utils up, moved AnimatePlacement.kt to core
