+++ features/src/main/kotlin/ru/mobileup/template/features/cryptocurrency/presentation/list/ui/CoinListToolbar.kt
@@ -17,10 +17,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import ru.mobileup.template.features.cryptocurrency.presentation.shared_ui.toolbars.ToolbarText
+import ru.mobileup.template.features.cryptocurrency.presentation.widget.toolbars.ToolbarText
 import ru.mobileup.template.core.theme.AppTheme
 import ru.mobileup.template.features.R
-import ru.mobileup.template.features.cryptocurrency.presentation.shared_ui.toolbars.AppToolbar
+import ru.mobileup.template.features.cryptocurrency.presentation.widget.toolbars.AppToolbar
 
 @Composable
 fun CoinListToolbar(
