+++ features/src/main/kotlin/ru/mobileup/template/features/cryptocurrency/presentation/utils/CurrencySymbol.kt
@@ -1,4 +1,4 @@
-package ru.mobileup.template.features.cryptocurrency.presentation.shared_ui.utils
+package ru.mobileup.template.features.cryptocurrency.presentation.utils
 
 import java.util.Currency
