+++ features/src/main/kotlin/ru/mobileup/template/features/cryptocurrency/presentation/widget/toolbars/ToolbarText.kt
@@ -1,4 +1,4 @@
-package ru.mobileup.template.features.cryptocurrency.presentation.shared_ui.toolbars
+package ru.mobileup.template.features.cryptocurrency.presentation.widget.toolbars
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
